[#1087] Split logic of CommentLine into a separate component and fixed OverviewLine bug.
